usb: move USB interface related classes to a new 'usb' package
Rename also the 'AndroidXBeeInterface' class to 'AndroidUSBInterface'.
