+++ xbee_android_library/src/test/java/com/digi/xbee/api/android/connection/android/CircularByteBufferReadTest.java
@@ -15,7 +15,7 @@
  */
 package com.digi.xbee.api.android.connection.android;
 
-import com.digi.xbee.api.android.connection.CircularByteBuffer;
+import com.digi.xbee.api.android.connection.usb.CircularByteBuffer;
 
 import org.junit.Before;
 import org.junit.Test;
